fix more signed/unsigned warnings
git-svn-id: http://svn.boost.org/svn/boost/trunk@42642 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/accumulators/statistics/peaks_over_threshold.hpp
@@ -235,7 +235,7 @@ namespace impl
                 );
 
                 // If n is in a valid range, return result, otherwise return NaN or throw exception
-                if ( n >= tail(args).size())
+                if ( n >= static_cast<std::size_t>(tail(args).size()))
                 {
                     if (std::numeric_limits<float_type>::has_quiet_NaN)
                     {
