+++ boost/accumulators/statistics/weighted_tail_mean.hpp
@@ -94,7 +94,7 @@ namespace impl
 
             while (sum < threshold)
             {
-                if (n < tail_weights(args).size())
+                if (n < static_cast<std::size_t>(tail_weights(args).size()))
                 {
                     sum += *(tail_weights(args).begin() + n);
                     n++;
